* README.sun: Added an explicit reminder at the top that the information in this file is historical in nature and does not necessarily apply to current reality.
* README: Updated the season from "summer" to "Fall" ;-) Also
added a note that warned people not to try to link TAO with a
version of ACE that's linked with another ORB library, e.g.,
liborbix.so.  Thanks to Steven Wohlever <wohlever@mitre.org> for
bringing this instructional omission to our attention.

* tao/tao_internals.cpp (fake_service_entries_i): Made the body of
this function conditional on the TAO_PLATFORM_SVC_CONF_FILE_NOTSUP
preprocessor definition since it doesn't need to exist in
platforms that grok the Service Configurator.  This will also
reduce the memory footprint :-).  Thanks to David Miron
<dxm@crapper.dsto.defence.gov.au> for pointing this out!

* tao/orb_core.cpp (TAO_ORB_Core): Removed explicit CTOR
initialization of reactor_ member.


git-svn-id: svn://svn.dre.vanderbilt.edu/DOC/Middleware/trunk@4990 ae88bc3d-4319-0410-8dbf-d08b4c9d3795